refactor: Consolidate styles into single file
- Merge colors.go, symbols.go, styles.go, and lipbalm_styles.go into output/styles.go
- Remove styles/ subdirectory (4 files → 1 file)
- Update imports and function calls to use local functions
- Better organization with all styles in one place
